lib/, src/: Remove useless casts in fgets(3)
This patch can be replicated with the following semantic patch:

	$ cat fgets_cast.sp
	@@
	expression a, b, c;
	@@

	- fgets(a, (int) (b), c)
	+ fgets(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgets(a, (int) b, c)
	+ fgets(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgetsx(a, (int) (b), c)
	+ fgetsx(a, b, c)

	@@
	expression a, b, c;
	@@

	- fgetsx(a, (int) b, c)
	+ fgetsx(a, b, c)

	@@
	expression a, b, c, p;
	@@

	- p->cio_fgets(a, (int) (b), c)
	+ p->cio_fgets(a, b, c)

	@@
	expression a, b, c, p;
	@@

	- p->cio_fgets(a, (int) b, c)
	+ p->cio_fgets(a, b, c)

which is applied as:

	$ find lib* src/ -type f \
	| xargs spatch --sp-file ~/tmp/spatch/fgets_cast.sp --in-place;

+++ lib/shadow/gshadow/fgetsgent.c
@@ -66,9 +66,7 @@ fgetsgent(FILE *fp)
 		buflen *= 2;
 
 		len = strlen (buf);
-		if (fgetsx (&buf[len],
-			    (int) (buflen - len),
-			    fp) != &buf[len]) {
+		if (fgetsx(&buf[len], buflen - len, fp) != &buf[len]) {
 			return NULL;
 		}
 	}
